chore: new auth screens ui
