+++ apps/app/components/auth-screens/index.ts
@@ -1,1 +1,3 @@
+export * from "./project";
 export * from "./workspace";
+export * from "./not-authorized-view";
