+++ apps/app/components/auth-screens/not-authorized-view.tsx
@@ -7,16 +7,16 @@ import { useRouter } from "next/router";
 import DefaultLayout from "layouts/default-layout";
 // hooks
 import useUser from "hooks/use-user";
-// img
-import ProjectSettingImg from "public/project-setting.svg";
+// images
+import ProjectNotAuthorizedImg from "public/auth/project-not-authorized.svg";
+import WorkspaceNotAuthorizedImg from "public/auth/workspace-not-authorized.svg";
 
-type TNotAuthorizedViewProps = {
+type Props = {
   actionButton?: React.ReactNode;
+  type: "project" | "workspace";
 };
 
-export const NotAuthorizedView: React.FC<TNotAuthorizedViewProps> = (props) => {
-  const { actionButton } = props;
-
+export const NotAuthorizedView: React.FC<Props> = ({ actionButton, type }) => {
   const { user } = useUser();
   const { asPath: currentPath } = useRouter();
 
@@ -29,7 +29,12 @@ export const NotAuthorizedView: React.FC<TNotAuthorizedViewProps> = (props) => {
     >
       <div className="flex h-full w-full flex-col items-center justify-center gap-y-5 text-center">
         <div className="h-44 w-72">
-          <Image src={ProjectSettingImg} height="176" width="288" alt="ProjectSettingImg" />
+          <Image
+            src={type === "project" ? ProjectNotAuthorizedImg : WorkspaceNotAuthorizedImg}
+            height="176"
+            width="288"
+            alt="ProjectSettingImg"
+          />
         </div>
         <h1 className="text-xl font-medium text-gray-900">
           Oops! You are not authorized to view this page
@@ -38,7 +43,7 @@ export const NotAuthorizedView: React.FC<TNotAuthorizedViewProps> = (props) => {
         <div className="w-full text-base text-gray-500 max-w-md ">
           {user ? (
             <p className="">
-              You have signed in as {user.email}.{" "}
+              You have signed in as {user.email}. <br />
               <Link href={`/signin?next=${currentPath}`}>
                 <a className="text-gray-900 font-medium">Sign in</a>
               </Link>{" "}
