+++ apps/app/components/auth-screens/project/index.ts
@@ -1,0 +1,1 @@
+export * from "./join-project";
