+++ apps/app/components/project/index.ts
@@ -1,6 +1,5 @@
 export * from "./create-project-modal";
 export * from "./delete-project-modal";
-export * from "./join-project";
 export * from "./sidebar-list";
 export * from "./single-integration-card";
 export * from "./single-project-card";
